Move the json to tests
